Made changes to profile and project pages and added images
Completed the media query of the profile page, added all my projects to the projects page and also completed the media query for the projects page, added social media icons to the footer. I sorted the images per folder inside of the image folder, one for socials, one for programming languages and one for extra pictures that don't have a specific category. I sorted the colors inside of root, making them more readable.