Made the skills page and a small profile css change
I created the skills page of my portfolio website, added all the skills that I know of like the programming languages which I know and also the actual languages that I know like Azerbaijani, English, etc. I also made a small change to the root of my profile css, there was a color there that wasn't being used so I removed it.